refactor(web): split host shell api
